Decompose cli a bit
